libusb NaCl port: fix issues with transfers cancellation, add tests
The issue that this commit is solving was losing of some input transfers
data, in cases when a transfer with the same parameters was previously
canceled. It was caused by a wrong assumption from chrome.usb API (which, by
itself, provides no means of cancelling a transfer) that all of the
simultaneously working input transfers will receive the data sent from the
device. When the chrome.usb response data arrived to a previously canceled
transfer, this data was simply discarded.

The external effect was that PC/SC-Lite daemon inside the Connector App could
report a reader as having a smart card inserted, while it's not, or vice
versa.

The fix for this issue involves restructuring the libusb NaCl port code. Now
the input transfers are handled differently: the link between libusb transfer
and the submitted chrome.usb transfer is not kept anymore; instead, the
received data is routed to a transfer according to the transfer parameters.
This allows to pass the data, that was originally intended to be passed to a
canceled transfer, to some another transfer in the future.

This solution is a compromise, because the cancellation of the input
transfers is still just an emulation, though more robust now.

Cancellation of output transfers is not supported anymore.