Avoid dying when chrome.certificateProvider API is unavailable in Release mode
The previous behavior was that the fatal failure occured in Release mode
when the chrome.certificateProvider was not found. This leads to the
restart of the whole app, which happens in an infinite loop until Chrome
finally stops the app completely.

As there is nothing useful about such behavior, it's now changed just to
emit a fatal error message and keep working (without integration with
chrome.certificateProvider API).