Filter out invalid characters from created Pepper values
This fixes the problem when log messages containing invalid (at least
non-ASCII) characters could not be delivered to the JS side, raising the
assertion errors instead (because the invalid strings are getting replaced
by null).

After this change, all non-printable characters (for the reference, see
<http://www.cplusplus.com/reference/cctype/isprint/>) in the log messages
are replaced with spaces before being sent to the JS side.

Also an assertion was added that no other code tries to pass non-printable
characters to the Pepper values construction helpers.