Fix crash caused by releasing a context without disconnecting
The crash could happen as a result of situation when a client
establishes a context, creates a handle after connecting to the card,
and then releases the context without disconnecting.

The PC/SC-Lite behavior in that case is that all handles to the
connected cards are automatically destroyed (i.e., essentially, the
disconnection happens).

However, the NaCl port code contains several security checks that
strictly control and validate all handles. So the crash happened when
PC/SC-Lite decides to re-use the the automatically freed handles, which
was unexpected by the NaCl port code and resulted in a fatal error.

So the fix is to make the NaCl port code aware of such situations and mark
all left handles as free when the corresponding context is destroyed.