Protect debug dumping code against undefined classes (#31)
This hopefully fixes an intermittent problem that was observed with debug dump function: it failed sometimes with TypeError saying that right-hand argument of "instanceof" operator cannot be undefined. The guess is that some of the referenced classes are not initialized by the time the extension code starts. So the commit checks whether the classes themselves evaluate to true before executing instanceof against them.