Remove extra logs that happened for each PC/SC client request
This commit removes in most cases the logging on the JavaScript side
(the ClientHandler class) that happened on each PC/SC client request.

Those logs were extra in most cases, as once the request reaches the
PC/SC-Lite server in the NaCl module, it will itself produce the
informative log messages.

So this commit removes this extra logging. Instead, a new logging is
added for the situations when there is a delay in processing because
of waiting for the PC/SC-Lite server initialization.