Fix wrong attribution of log messages in the export
This commit fixes the issue with the logs export with attributing
all emitted log messages to the background page regardless of which
window they came from.

Additionally, this commit introduces pre-formatting of the stored
log messages, which makes the log buffer consume less memory and makes
the log exporting faster.