Display a warning when running Connector on non-Chrome OS system (#35)
This commit adds a warning message displayed in the Connector app's window when it's executed on a non-Chrome OS platform. The reason for this is that, on one side, we want to discourage people from running it on the desktop platforms (because it can only bring very limited benefit on them); while on the other side, we want to allow the Connector app to be used by third-party client app developers for testing purposes.

Fixes #27.