Respond with a ping to received ping immediately when in single-message mode
This is workaround for client inability to tell the Connector app that it
was reloaded, when the single-message-based communication is used. With this
change, the Connector will send an additional (out-of-the-schedule) ping
message to the client each time it receives a ping message from a client.

Before this change, if a client was reloaded between two ping requests from
the Connector app, it could send a bunch of commands before the ping request
comes from the Connector app. After the Connector app receives the response for
its new ping request and realizes that the client was reloaded, it would
clean up all data that was allocated before that. As a result, this puts the
client into a weird situation without good ability to tell the Connector app
to start a clean session of commands.