Persist channel id across client reconnections
This commit makes the "channel_id" key value in the "pong" messages be
effectively constant during the whole lifetime of the app.

This fixes the problem with the restarted clients, when they could
receive different responses for the first "ping" request and for some
subsequent "ping" requests, because previously the channel id was
generated from scratch each time a client connects, and the detection
of the client reloading happens usually only after the first "ping"
request was processed by the Connector app.