Fix ignoring the first message from a reloaded client
This commit fixes the problem with losing the first message from a
reloaded client when the single-message-based communication is used.

The problem was caused by not re-creating the message channel
immediately after the client reload was detected. As a result, the
Connector forgets about the client until the new message from it is
received, and the message would be considered as not established
until a new "pong" respond would be received from the client. This
situation could cause a significant delay in processing client's
requests or even not processing the client's requests at all
(depending on how the client is implemented).