Cancel blocking PC/SC calls on client disconnection.
This fixes a problem with the Connector app that it doesn't release the
handles obtained by a client when the client is disconnected during a
blocking PC/SC call.

The problem could result, for instance, in blocking (temporarily or even
forever) the requests from a quickly restarted client.

This closes #43.