Auto-generate Connector app's manifest from a template
Before this commit, the Connector app's manifest.json file had to be updated
manually together with each CCID update, as the list of supported USB devices
in the manifest has to correspond to the list of smart card readers supported
by CCID.

This commit adds a script and makefile rules for automatic generation of the
manifest file from a template and from the CCID supported readers config file.
This closes #39.

Also this commit clears the road for the future enhancements of the development
process, which could now be based on using modified versions of the manifest
file.