Fix C++ client example to handle early received certificateProvider API events
This fixes the issue in the C++ client example code, where an event received
from the chrome.certificateProvider API before the NaCl module has loaded was
ignored. As Chrome considers not handling the event as an error, this could
result in bad UX in cases when the navigation occurs before the NaCl module
loads (which could take quite some time on non-top Chromebooks).

This commit adds buffering of the received events into the
CertificateProviderBridge.Backend JS class, which allows to set up event
handlers straight away when the app's background script starts.

This closes #37.