Fix JS error in requester after disposal
The JavaScript SyntaxError could be triggered in the
Requester.responseMessageReceivedListener_ method if the requester
has already been disposed by that point.

This could happen in various scenarios in form of a race between the
requester disposal and receiving a response through a message channel.
One example of such scenario is a frequently connecting and
disconnecting client.