Show user prompt even for requests from unknown apps
This change relaxes the security checks in the Smart Card Connector
app, to make the development and testing against it simpler.

The Connector app will now show user prompts even for the unknown
client apps (i.e. the apps whose ID's are not in the
known_client_apps.json whitelist).

Previously, the user prompts from unknown client apps were silently
ignored without any UI feedback.

Note that after this change, a big scary warning message will be
displayed when asking about unknown apps, as this feature is
primarily oriented to developers and testers and not to the end
users.