Fix SCardTransmit returning SCARD_PROTOCOL_ANY as the receive protocol
The NaCl port of the SCardTransmit function always returns the receive
protocol, even when it was not supplied by the caller. However, it was
returning the value 3 (that is, SCARD_PROTOCOL_ANY) is such cases,
which is not quite correct.

This fixes this by replacing this placeholder value with the value of
the send protocol. This replacement seems to be correct in most cases.
And it's anyway not possible to obtain the actually used receive
protocol in the NaCl port, due to how PC/SC-Lite and CCID are
currently implemented.

This closes #51.