Display successfully initialized readers on top
This changes the sorting order in the list of the readers in the Smart
Card Connector app's window.

Previously, it displayed failed readers on top of the successfully
initialized readers, which could lead to user confusion - especially in
case of composite devices. It could be particularly weird if the number
of displayed items exceeds 2 - then the last items with the successfully
initialized readers will become hidden, and only available by scrolling
(with a tiny scroll bar).

This fixes #52.