Heuristics for hiding useless errors from composite devices
This change attempts to hide from Smart Card Connector app's UI items
that tell an error, but don't actually correspond to a real error
with a smart card reader. This happens with composite devices, where
previously each interface of them was represented with a separate
item in the UI - which was always a failure item for non-reader
interfaces.

The heuristics is to parse the device name, as reported by PC/SC-Lite
server, and filter out all failed devices which have a non-zero
interface in their name. This is expected to work, as CCID in
typical environment anyway ignores the interface numbers, and
attempts to guess the right interface to talk to with the first
supplied item (which has the interface equal to zero in its name).

This is follow up to issue #52.