Add test-all.sh script for running all tests in repo
This script simplifies execution of all tests in the project
repository. It hardcodes paths to all required targets, which removes
possibility of forgetting about some target when typing this all
manually.

Additionally, it executes all tests in both Debug and Release
configurations.

The major problem with the script, however, is that the tests
themselves are still not fully automated: someone has to look into the
opened browser windows and check that all tests pass. Moreover, the
testing doesn't proceed to the next target before the human closes the
window with the previous test target execution.