Import Closure Compiler sources, remove Chrome
Import the Closure Compiler sources 20180101, in addition to the
previously used import of the binary build. This allows to use the
extern definitions for Chrome APIs.

One reason why this became necessary is some recent change in the
Closure Compiler that it doesn't make standard JS APIs' externs
available by default. Another reason is that this allows to remove the
(partial) checkout from the Chromium repository.