Fix incorrect storage of user selection data
The Connector app was incorrectly serializing the user selection data
when saving it in the local storage, in cases when the stored data
contained more than one item.

Besides corrupting the stored user selections, this bug was leading to
the app crashes and crash-restart loops, as the error was unrecoverable.

This fixes #57.