Update cacerts.txt in NaCl SDK
Update the cacerts.txt file in the NaCl SDK to contain an
up-to-date certificate that is required in order to talk to
storage.googleapis.com - the "GlobalSign_Root_CA_-_R2"
certificate.

This is needed for addressing #71, since currently the NaCl SDK
cannot fetch the Pepper SDK due to TLS error when doing network
requests.