Wrap Chrome --pack-extension run into Xvfb
Run the Chrome binary, in case when it's called by the build scripts
just to produce a CRX file and not to draw any UI, via the xvfb-run
binary.

This allows to build the CRX files on systems that don't have an X11
server running (e.g., remote virtual machines).

Without this change, Chrome fails with an error like this without
producing the CRX file:

  Gtk-WARNING **: 12:00:00.000: cannot open display