C++ example: Use requestPin API provided by Chrome
Change the example_cpp_smart_card_client_app skeleton to use the
chrome.certificateProvider.requestPin API instead of the
customly crafted HTML popup.

This API has been the recommended approach since Chrome 57. Its
benefits are: the UI aligned with the rest of Chrome OS, ability
for Chrome to show the PIN dialog associated with the context that
triggered it, ability to work on Chrome OS login screen.