Best-effort fix for NaCl crash log message delivery
Wait a little bit between emitting the FATAL log message and
crashing, so that the emitted log message has a chance to be
delivered to the JavaScript side where it's easier to be
discovered by the developer.

This fix is imperfect and works only with some probability
(depending on the system load and, likely, on whether the fatal
error happened on the main thread or on a background thread). But
that's still an improvement, and also the error logging in this
case is best-effort anyway.