Allow logging the PC/SC-Lite API call results
Add the implementation of logging the results of the PC/SC-Lite API
calls that are made from a NaCl module. This should aid in the
future in the investigation of bugs in third-party middleware
extensions.

The implementation is disabled by default, since even under normal
circumstances a typical smart card middleware makes dozens of
PC/SC-Lite requests per single certificates listing / signature
operation.

This additional logging can be enabled by running these snippets in
the JavaScript console of the extension's background page:

* enabling the logging of return codes:

  $jscomp.scope.pcscLiteNaclClientBackend.logger.setLevel(goog.log.Level.FINE)

* additionally, enabling the dumping of all values and buffers (use
  with care - this may include privacy/security-sensitive data) :

  goog.DEBUG=true