Enable unlocking user profile from lock screen. Close #100. Close #90.
Fix bug #100: Previously, smartcard users could not unlock the lock screen since
the user space smartcard connector app would still own the smartcard reader.
This makes the app show no available devices when the user is locking the
screen. PCSC automatically releases devices that it used to own but no longer
show as present.

This alone is not sufficient to solve the problem. PCSC scans devices only once
every second. When locking the screen, the login-space app might already try to
access the reader before the user-space app releases it. This race condition is
fixed by issue #90: When connecting to a reader fails, we retry connecting
later. This is done by showing another Bus number for the device.
We need to fake Bus numbers anyway as chrome.usb does not let us retrieve them.
PCSC then views it as a different device from the previous one and attempts
to connect to it once a smartcard reader needs to be accessed.