Avoid dumping non-struct classes for logging
Add a heuristic check that omits performing recursive dumping for
objects that inherit from goog.Disposable, since such objects are most
likely classes that aren't useful to be logged (as they are not pure
containers for data) and are expensive to be logged (as they may contain
non-trivial state).

Also, in accordance with this heuristic, make the
GoogleSmartCard.LogBuffer class inherit from goog.Disposable - since
objects of this class are one of the most heavy-weight objects in the
whole extension, and we definitely want to avoid it being debug-dumped
at any time.