C++ example: Restore the built-in PIN dialog
Resurrect the code that allows to show the PIN dialog as an HTML window
ordered by the App.

This PIN dialog is not recommended for general use, however it's still
useful in some scenarios - like requesting the PIN during the
"demo"/"test".

This change is a partial revert of commit
045219c11f7e191215dd2b177e51cb5983688c66.