Fix PIP installation in webports
Work around the issue with PIP installation during "webports" "gclient
sync" command, which manifests itself as import failures when running
PIP despite it being installed just before this command.

The hack is to unset the "PYTHONNOUSERSITE" environment variable which
is set by gclient/vpython: for some reason, having this variable set
makes PIP being not found.

This is a quick hack solution, in order to unblock developers - an ideal
solution would be to find out what's wrong in the way how we install
PIP.

This fixes #115.