Fix cert validation failures in NaCl SDK initialization (#124)
Fix for the nacl_sdk initialization issue that happen on some machines
like Ubuntu 20.04:

  "Host storage.googleapis.com returned an invalid certificate"

The problem was that the networking code in the NaCl SDK
initialization Python scripts was relying on TLS 1.0, meanwhile this
protocol has been deprecated a long time ago, and the modern
Linux distributions started to disable it in the OpenSSL
configuration.

The fix is to get rid of the legacy "fancy_urllib" library which only
worked with TLS 1.0, and use the standard Python networking
libraries that should work with any protocol negotiated between
the server and the client OpenSSL library. Also stop using and
delete the "cacerts.txt" file and rely on the default trust roots
installed into the system.