Extract PC/SC-based reader status tracker
Move the JS code that allows to track the readers state from the
server's internal into the client directory, which allows it to be
reused in client or middleware extensions.

No functional changes were made in that code - only the internal data
format was changed (to avoid having a structure with unused fields) and
the trivial "var ==> let/const" modernization was performed.