Omit "<background page>." from logs
Don't mention the background page in the JavaScript/NaCl log messages -
this information can be inferred from the absence of the HTML page URL.
Given that the majority of log messages come from the background page,
it makes sense to reduce the log clutter (and the RAM usage) by omitting
this part.

This change is part of the effort tracked by issue #146.