Fix minification issue in LogBufferForwarder (#153)
Fix a potential problem in LogBufferForwarder that the
"formatted_log_message" object key could be minified by the Closure
Compiler to something like "a", in case the aggressive optimization is
enabled.

For the background, the Closure Compiler in the aggressive compilation
mode renames all object properties to some minified names. This is
unwanted in this case, since the object constructed here is sent to the
NaCl module where it gets parsed (and the keys are expected to have
specific names). The way to suppress the Closure Compiler optimizations
is to put the property name into quotes.

This is a follow-up to #105 where this code was originally introduced.