Call runtime.reload() before runtime.restart()
Both functions achieve the same goal - reloading the extension - but
they are supported in different contexts: one only works in non-kiosk
sessions and another only in kiosk ones. We call both of them, in order
to simplify the code.

The commit swaps the order of these calls, so that reload() is called
first. This allows to remove the spurious error from restart() when run
in non-kiosk mode.

This is part of the crash-and-reload process improvement tracked by #156.