Add flag for disabling the reload-on-failure (#162)
Add a build-time parameter that allows to enable/disable the
self-reloading of the extension/app in case a fatal error occurs. This
allows to easily disable this functionality in client applications that
don't want this behavior.

Note that this parameter also controls whether the "crash loop
detection" introduced in #156 is enabled.

Also leave this parameter in the "disabled" state for the
"example_js_standalone_smart_card_client_library", since this library is
used in a wide variety of different extensions/apps, and, most likely,
they don't want to be self-reloaded in case the smart card code fails.
It was an oversight that this functionality was enabled there, and now
with addition in #156 of chrome.storage manipulations it became more
apparent that we want to disable it there by default.