Fix type definition in cpp example app
Fix definition of ClientCertificateInfo, which uses a vector of vectors
instead of a simple vector.
The inner vector is wrapped in a struct, since otherwise pepper fails to
extract the value.
Also add a note that the outer vector, right now, cannot contain
more than one element as per API definition.