Omit '<"nacl_module.nmf">.NaCl' from logs (#172)
Don't adding the '<"nacl_module.nmf">.NaCl' suffix to the logger name
used for NaCl-related log messages. This information about the URL of
the NaCl module is redundant - we don't use multiple NaCl modules
currently.

For the potential case when the detailed logging might be needed, a
boolean flag is kept in the code, allowing to revert to the old
behavior. Currently the flag is always set to false.

This change is part of the effort tracked by issue #146.