Framework for JS-and-C++ integration tests
Implement a mini-framework that allows to implement hybrid tests that
exercise both JavaScript and C++ code. In such tests, the main test
script will be written in JavaScript, meanwhile the C++ counterpart is
structured as a set of "helpers" that can be called from the JS code.
The implemented framework abstracts away some boilerplate related to
JS-and-C++ communication and code compilation.

This commit paves the way for fixing the missing test coverage for the
parts that cross the language boundary. Up to this point, we only had
unit test coverage for pieces within single language (only within
JavaScript or only within C/C++), which has been insufficient.