Skip existing files/dirs in env/initialize.sh
Change the env/initialize.sh script to be idempotent by default, so that
already existing tools and libraries are left intact by it. Add the "-f"
option that allows to revert to the previous "destructive" behavior.

This is a preparation change for introducing the WebAssembly build tools
initialization in #177, since it'll make it possible for developers to
rerun "initialize.sh" in order to get WebAssembly tools without
destroying their existing Native Client related state.