Download Emscripten in env/initialize.sh (#180)
Add download and installation of Emscripten into the env/initialize.sh
script. This will be used in the future for building the WebAssembly
ports of the C/C++ code in the Smart Card Connector app and related
projects (instead of the currently used Native Client).

The version of Emscripten is pinned, so that all developers get
consistent results and avoid any potential breakages related to uprevs.