Introduce variant Value type
Initial implementation of the "Value" class that, approximately, can
hold any JSONifiable object.

This class will be used in follow-ups to replace the widespread usage of
the Native Client / Pepper specific "pp::Var" class, so that all
depending code can be reused in WebAssembly builds.

The class is additionally inspired by the Chromium's "base::Value", but
we're having a much simpler custom implementation here, because
Chromium's implementation brings a bunch of dependencies and is not
trivial to compile.