Rename NACL_MODULE_BUILD_RULE Makefile rule
Rename the "NACL_MODULE_BUILD_RULE" Makefile rule into
"LINK_EXECUTABLE_RULE".

This is part of the WebAssembly migration effort (as tracked by #177) -
a generic name in this macro allows to use it in a toolchain-agnostic
way, so that only its implementation is aware of "Native Client vs.
WebAssembly" differences.