Use StructConverter for typed messages
Refactor the code that creates/parses "typed message" dictionaries to
hold the data in a dedicated struct and to use the StructConverter for
conversion to/from pp::Var instances.

This refactoring removes one particular usage of the
"VarDictValuesExtractor" converter class; the goal is to get rid of all
usages of all usages of this class except the StructConverter itself.
This will make it easier to migrate to the Value class in the future
(which is part of the WebAssembly migration, as tracked by #185),
since the Value-based conversions will use the interface similar to
StructConverter.