Add optional::reset() method (#190)
Extend our polyfill for the "optional" C++ class to have the reset()
method, and use it in places where the optional was reset by other
means.

Historically, the original C++ Standard proposal for the std::optional
type didn't have the reset() method, which is why we didn't add into our
polyfill as well. However, afterwards this method was added into the
final std::optional interface, which went into C++17.

In the future, once we fully migrate to WebAssembly and drop the Google
Native Client support, we'll be able to delete the whole polyfill class
and use the "optional" class provided by the C++ Standard Library.