Migrate MessageListener from pp::Var to Value (#196)
Change the MessageListener C++ interface and all its implementations to
stop using the Native Client's pp::Var class, and use the
toolchain-independent Value class.

This is part of the effort on adding WebAssembly support into
//common/cpp (as tracked by #185). This MessageListener class is,
effectively, an entry point into the C/C++ code of incoming JavaScript
message handling. It's called by toolchain-specific code (currently,
only Native Client specific wrappers, but in the future an alternative
WebAssembly specific wrapper will be added). This commit only affects
one of the first steps of the pipeline - subsequent ones will be done in
follow-up commits.

As a bonus, this commit is making error messages more detailed in
case when the incoming JS message handling failed.