Add identity transform for Value into itself
Add overloads for the ConvertToValue/ConvertFromValue functions that
perform identity transform, i.e., return the input argument without
changes.

This is needed in order to make generic transformation code work in case
when the same Value class is used in both source and destination
representation - for example, the TypedMessage struct's |data| field.