Add "OrDie" Value conversion variants
Add new variants of the ConvertToValue/ConvertFromValue functions that
immediately crash the program in case the conversion fails.

These functions should be used in cases when the error is really
unexpected (e.g., when the input data is guaranteed to be of specific
form) and/or when there's no recovery possible from the error.