Migrate typed_message.h to Value
Change the typed_message.h and the corresponding .cc file to use the
Value class and drop all usages of Native Client specific pp::Var
classes from it.

This removes a bunch of boilerplate related to the struct-to-pp::Var and
inverse conversions. Consequently, the unit tests are deleted, since
there's nothing to test here anymore (conversions between structs and
Value objects are covered by a generic unit test).

This commit contributes to the goal of making most of the //common/cpp
library be toolchain-agnostic and support WebAssembly (as tracked
by #185).