Script for auto-formatting code via clang-format
Add a small script that can be used when developing locally in order to
automatically format the touched C++ files according to the Google C++
Style Guide. The script is based on the "clang-format" tool.

The script assumes that the "clang-format" program is installed into the
system and is in $PATH.