Migrate non-//common to Value-based TypedMessage
Adjust the C++ code outside //common/cpp to the new implementation of
TypedMessage that doesn't support pp::Var anymore.

This CL is a stop-gap solution to make the whole project still compile
and work while we're working on the migration of the //common/cpp
library (as tracked by #185). After completion of that part, we'll clean
up the non-//common code to delete all those back-and-forth pp::Var
conversions and use Value all the time.