Split out toolchain-specific .mk build rules
Add single nacl_module_building_common.mk file that provides a
toolchain-independent interface for building C/C++ code, and move the
Native Client specific implementation of this interface into
internal/binary_executable_building_nacl.mk.

This structure is one of the steps of adding the support of WebAssembly
builds (as tracked by #177) while keeping the Native Client support for
the transition period.

This is a mechanical change, except minor changes in the build rules
documentation.