Migrate typed_message_listener.h to Value
Change the typed_message_listener.h and all child classes to use the
Value class in the interface, rather than using the Native Client
specific pp::Var.

As an additional outcome, the implementation of the
typed_message_router.cc file becomes completely free of pp::Var usages
as well.

This commit contributes to the goal of making most of the //common/cpp
library be toolchain-agnostic and support WebAssembly (as tracked
by #185).