Migrate requester_message.h to Value (#206)
Change the requester_message.h header to use the toolchain-independent
Value class instead of Native Client specific pp::Var class. Besides the
mechanical replacement, this commit introduces dedicated structs for the
"request"/"response" messages, since it makes the interface more concise
and allows to reuse automatic to/from Value conversion helpers.

This commit contributes to the goal of making most of the //common/cpp
library be toolchain-agnostic and support WebAssembly (as tracked
by #185).