Migrate [js_]requester.h to Value
Change requester.h and js_requester.h headers and the corresponding
implementation .cc files to use the toolchain-independent Value
class intead of Native Client specific pp::Var class in most places.
The only left place is the delegate for sending messages to JS, which
will be refactored separately.

This commit contributes to the goal of making most of the //common/cpp
library be toolchain-agnostic and support WebAssembly (as tracked
by #185).