Migrate remote_call_message.{h|cc} to Value
Change the remote_call_message.h and .cc files to use the
toolchain-independent Value class instead of the Native Client specific
pp::Var class. Brief overview of changes in this commit:

* Introduce a dedicated RemoteCallRequestPayload struct for holding all
  fields of the message's payload - this allows to reuse the automatic
  to/from Value conversion mechanism and reduces the boilerplate when
  passing this data around.
* Implement the "any-to-RemoteCallRequestPayload" conversion. (The old
  NaCl-specific analog was MakeVarArray() from
  pp_var_utils/construction.h.)
  This conversion is covered with its own unit tests.
* Mechanical replacements to support the new interface.

This commit contributes to the goal of making most of the //common/cpp
library be toolchain-agnostic and support WebAssembly (as tracked
by #185).