Introduce interface for toolchain-specific C++ ops
Introduce a GlobalContext interface that abstracts away some of the
differences between Native Client and Emscripten, such as different ways
of sending messages to JavaScript.

This commit provides implementations of this interface for both Native
Client and Emscripten. Currently both the interface and the
implementations are unused; follow-up commits will migrate the existing
codebase in //common/cpp to use this interface instead of various
NaCl-specific delegates (contributing to the goal of making most of the
//common/cpp library be toolchain-independent, as tracked by #185).